creating primary view js
